--- src/main/kotlin/co/ao/nextbss/requirekt/CustomErrorResponse.kt
@@ -1,8 +1,0 @@
-package co.ao.nextbss.requirekt
-
-import co.ao.nextbss.requirekt.annotation.CustomErrorResponse
-
-@CustomErrorResponse
-interface CustomErrorResponse {
-    fun toJSON(): String
-}
