+++ src/main/kotlin/co/ao/nextbss/requirekt/ErrorResponse.kt
@@ -1,0 +1,8 @@
+package co.ao.nextbss.requirekt
+
+import co.ao.nextbss.requirekt.annotation.ErrorResponse
+
+@ErrorResponse
+interface ErrorResponse {
+    fun toJSON(): String
+}
